# docs/swagger.yaml
basePath: /api
definitions:
  github_com_eminbekov_fiber-v3-template_internal_dto_request.CreateUserRequest:
    properties:
      full_name:
        example: John Doe
        maxLength: 100
        type: string
      password:
        example: StrongPass123!
        minLength: 8
        type: string
      phone:
        example: "+998901234567"
        type: string
      username:
        example: john_doe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - full_name
    - password
    - phone
    - username
    type: object
  github_com_eminbekov_fiber-v3-template_internal_dto_request.LoginRequest:
    properties:
      password:
        example: StrongPass123!
        type: string
      username:
        example: john_doe
        type: string
    required:
    - password
    - username
    type: object
  github_com_eminbekov_fiber-v3-template_internal_dto_request.UpdateUserRequest:
    properties:
      full_name:
        example: John Doe
        maxLength: 100
        type: string
      phone:
        example: "+998901234567"
        type: string
      status:
        enum:
        - active
        - disabled
        example: active
        type: string
      username:
        example: john_doe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - full_name
    - phone
    - status
    - username
    type: object
  github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorBody:
    properties:
      details:
        items:
          $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.FieldError'
        type: array
      message:
        example: validation failed
        type: string
    type: object
  github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorBody'
    type: object
  github_com_eminbekov_fiber-v3-template_internal_dto_response.FieldError:
    properties:
      field:
        example: username
        type: string
      message:
        example: username is required
        type: string
    type: object
  github_com_eminbekov_fiber-v3-template_internal_dto_response.Response:
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: REST API for authentication, authorization, and user management.
  title: Fiber v3 Template API
  version: 1.0.0
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates with username and password and returns a session
        token.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.Response'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidates the current authenticated session.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
  /v1/ping:
    get:
      consumes:
      - application/json
      description: Returns a simple payload to verify API availability.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.Response'
      summary: Ping API
      tags:
      - Health
  /v1/users:
    get:
      consumes:
      - application/json
      description: Returns users with page and page_size query parameters.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.Response'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user and returns the created user payload.
      parameters:
      - description: Create user request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.Response'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Users
  /v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes a user and returns no content on success.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Returns one user by UUID identifier.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.Response'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates user fields and returns the updated user payload.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.Response'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_eminbekov_fiber-v3-template_internal_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Enter Bearer token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"